remove new_fastring
