+++ pyat/at/physics/fastring.py
@@ -1,114 +1,200 @@
 """
-Functions relating to fast_ring
+Functions relating to fast_ring.
 """
 
 from __future__ import annotations
 
 __all__ = ["fast_ring", "simple_ring"]
 
-import copy
 from collections.abc import Sequence
-from functools import reduce
-
 import numpy as np
-
-from ..constants import clight
-from ..lattice import Lattice, Particle, Refpts
-from ..lattice import RFCavity, DeltaQ, Marker, get_cells, checkname
-from ..lattice import get_elements, M66, SimpleQuantDiff, AtError, SimpleRadiation
+from ..lattice import Lattice, Refpts, AtError, Particle
+from ..lattice import Drift, RFCavity, Element, Marker, SimpleQuantDiff
+from ..lattice import DeltaQ, SimpleRadiation, M66
 from ..physics import gen_m66_elem, gen_detuning_elem, gen_quantdiff_elem
+from ..constants import clight
 
 
-def _rearrange(ring: Lattice, split_inds: Refpts = None):
+def _replace_cav(cav_element: Element) -> Element:
+    name = ("DR_" + cav_element.FamName,)
+    elem = Drift(name, cav_element.Length) if cav_element.Length > 0 else Marker(name)
+    return elem
+
+
+def _merge_cavs(all_cavs: Sequence) -> Sequence[RFCavity]:
+    m_cavs = []
+    freqs = [e.Frequency for e in all_cavs]
+    for fr in np.atleast_1d(np.unique(freqs)):
+        cavf = [cav for cav in all_cavs if cav.Frequency == fr]
+        vol = np.sum([c.Voltage for c in cavf])
+        cavl = RFCavity(
+            cavf[0].FamName,
+            0,
+            vol,
+            fr,
+            cavf[0].HarmNumber,
+            cavf[0].Energy,
+            TimeLag=cavf[0].TimeLag,
+        )
+        m_cavs.append(cavl)
+    return m_cavs
+
+
+def _rearrange(ring) -> tuple:
+    cav_idx = ring.get_uint32_index(RFCavity)
+    mcavs = _merge_cavs(ring[cav_idx])
+    ring[cav_idx] = [_replace_cav(ring[i]) for i in cav_idx]
+    ring += mcavs
+    ring += [Marker("xsplit")]
+
+
+def _split_ring(
+    ring: Lattice, split_inds: Refpts | None = None, keep_cavities: bool = False
+) -> Sequence:
     inds = ring.get_bool_index(split_inds, endpoint=True)
-    inds[0] = True
-    inds[-1] = True
+    inds[[0, -1]] = True
+    if keep_cavities:
+        cavsb = ring.get_uint32_index(RFCavity)
+        cavse = cavsb + 1
+        inds = inds | ring.get_bool_index(cavsb) | ring.get_bool_index(cavse)
     inds = ring.get_uint32_index(inds)
-    all_rings = [ring[int(b) : int(e)] for b, e in zip(inds[:-1], inds[1:])]
-
-    ringm = []
-    for ring_slice in all_rings:
-        ring_slice.insert(0, Marker("xbeg"))
-        ring_slice.append(Marker("xend"))
-        cavs = [e for e in ring_slice if isinstance(e, RFCavity)]
-        newpass = ["IdentityPass" if c.Length == 0 else "DriftPass" for c in cavs]
-        for c, pm in zip(cavs, newpass):
-            c.PassMethod = pm
-        uni_freq = np.unique([e.Frequency for e in cavs])
-        for fr in np.atleast_1d(uni_freq):
-            cavf = [c for c in cavs if c.Frequency == fr]
-            vol = reduce(lambda x, y: x + y, (c.Voltage for c in cavf))
-            cavl = RFCavity("CAVL", 0, vol, fr, cavf[0].HarmNumber, cavf[0].Energy)
-            cavl.TimeLag = cavf[0].TimeLag
-            ring_slice.append(cavl)
-        ringm = ringm + ring_slice
-    return all_rings, Lattice(ringm, energy=ring.energy, periodicity=1)
+    split_ring = []
+    for b, e in zip(inds[:-1], inds[1:], strict=True):
+        split_ring += [Marker("xsplit")]
+        split_ring += ring[int(b) : int(e)]
+    split_ring += [Marker("xsplit")]
+    split_ring = Lattice(split_ring, **vars(ring))
+    if keep_cavities is False:
+        _rearrange(split_ring)
+    return split_ring, split_ring.get_uint32_index("xsplit")
 
 
-def _fring(ring, split_inds: Refpts = None, detuning_elem=None):
-    all_rings, merged_ring = _rearrange(ring, split_inds=split_inds)
-    ibegs = get_cells(merged_ring, checkname("xbeg"))
-    iends = get_cells(merged_ring, checkname("xend"))
-    _, orbit = merged_ring.find_orbit(refpts=ibegs | iends)
-    if detuning_elem is None:
-        detuning_elem = gen_detuning_elem(merged_ring, orbit[-1])
-    else:
-        detuning_elem.T1 = -orbit[-1]
-        detuning_elem.T2 = orbit[-1]
-
-    fastring = []
-    for counter, r in enumerate(all_rings):
-        cavs = [e for e in r if e.PassMethod.endswith("CavityPass")]
-        [r.remove(c) for c in cavs]
-        lin_elem = gen_m66_elem(r, orbit[2 * counter], orbit[2 * counter + 1])
-        lin_elem.FamName = lin_elem.FamName + "_" + str(counter)
-        [fastring.append(c) for c in cavs]
-        fastring.append(lin_elem)
-    fastring.append(detuning_elem)
-    try:
-        qd_elem = gen_quantdiff_elem(merged_ring)
-        fastring.append(qd_elem)
-    except ValueError:  # No synchrotron radiation => no diffusion element
-        pass
-    fastring = Lattice(fastring, **vars(ring))
-    return fastring
-
-
-def fast_ring(ring: Lattice, split_inds: Refpts = None) -> tuple[Lattice, Lattice]:
-    """Generates a "fast ring"
-
-    A fast ring consisting in:
+def fast_ring(
+    ring,
+    split_inds: Refpts | None = None,
+    qpx: Sequence[float] | None = None,
+    qpy: Sequence[float] | None = None,
+    detuning_coeff: Sequence[float] | None = None,
+    alphac: Sequence[float] | None = None,
+    emitx: float | None = None,
+    emity: float | None = None,
+    espread: float | None = None,
+    keep_cavities: bool = False,
+    dp: float | None = None,
+) -> Lattice:
+    """
+    A fast ring consisting in the following elements.
 
     * a RF cavity per distinct frequency,
     * a 6x6 linear transfer map,
     * a detuning and chromaticity element,
-    * a quantum diffusion element (for radiation ring).
+    * a quantum diffusion element active when radiations are enabled.
 
-
-    2 new lattices are returned, one with radiation and one without. These lattices
-    keep the same attributes (energy, particle, circumference, periodicity,…)
-    as the initial one.
+    A new lattices is returned with the same attributes (energy, particle,
+    circumference, periodicity,…) as the initial one. The radiation state is
+    also concerved. Radiations can be turned on or of using ring.disable_6d()
+    or ring.enable_6d(). *Rad attribute will be used when radiations are turned
+    on.
 
     It is possible to split the original ring in multiple "fastrings"
     using the ``split_inds`` argument.
 
+    By default all the cavities are merged ans placed at the end of the ring this
+    allows to minimize the number of segments and therefore lattice elements.
+    Howevever, the longitudinal dynamics is modified by this action. In order to
+    preserve the original RF configuration the ``keep_cavities`` argument should
+    be set to true. In this configuration the tracking is slower as more segment are
+    introduced.
+
     Parameters:
         ring:       Lattice description
         split_inds: List of indexes where to split the ring
+        qpx:            horizontal chromatic detuning coefficients. Default None.
+          If specified qpy should also be provided, if not first order term computed
+          automatically from ring. at.physics.chromaticity can be used to compute
+          higher order chromatic coefficients
+        qpy:            vertical chromatic detuning coefficients. Default None.
+          If specified qpy should also be provided, if not first order term computed
+          automatically from ring. at.physics.chromaticity can be used to compute
+          higher order chromatic coefficients
+        detuning_coeff: First order amplitude detuning coefficients
+          [dQx/dJx, dQx/dJy, dQy/dJy]. Default None: coefficient computer from ring.
+          at.physics.detuning can be used to compute detuning coefficients
+        alphac: Higher order momentum compaction factor. Default None. Can be
+          Can be claculated with physics.get_mcf()
+        emitx: Equilibrium horizontal emittance. Default is the one given by ring
+        emity: Equilibrium vertical emittance. Default is the one given by ring
+        espread: Equilibrium energy spread. Default is the one given by ring
+        keep_cavities: The ring will be sliced at each cavities. This allows to
+          match longitudinal dynamics of the original ring bu t slows down the
+          tracking as more segments are generated
 
     Returns:
-        fring (Lattice):    Fast ring without radiation
-        fringrad (Lattice): Fast ring with radiation
+        fastring (Lattice):    Fast ring lattice object
     """
-    ringi = ring.deepcopy()
-    fastringnorad = _fring(ringi.radiation_off(copy=True), split_inds=split_inds)
-    detuning_elem = copy.deepcopy(get_elements(fastringnorad, "NonLinear")[0])
-    fastringrad = _fring(
-        ringi.radiation_on(copy=True),
-        split_inds=split_inds,
-        detuning_elem=detuning_elem,
+    split_ring, split_inds = _split_ring(ring, split_inds, keep_cavities)
+    split_ring.set_rf_frequency(dp=dp)
+    _, o4 = split_ring.disable_6d(copy=True).find_orbit(refpts=split_inds, dp=dp)
+    _, o6 = split_ring.enable_6d(copy=True).find_orbit(refpts=split_inds, dp=dp)
+
+    fastring = []
+    for o4b, o4e, o6b, o6e, sib, sie in zip(
+        o4[:-1], o4[1:], o6[:-1], o6[1:], split_inds[:-1], split_inds[1:], strict=True
+    ):
+        r = split_ring[sib + 1 : sie]
+        iscav = np.array([isinstance(elem, RFCavity) for elem in r], dtype=bool)
+        if np.all(iscav):
+            fastring = [*fastring, *list(r)]
+        elif np.all(~iscav):
+            lin_elem = gen_m66_elem(
+                r.disable_6d(copy=True), o4b, o4e, r.enable_6d(copy=True), o6b, o6e
+            )
+            fastring = [*fastring, lin_elem]
+        else:
+            msg = (
+                "One fast ring segment contains a combination of magnets and cavity."
+                "They need to be separated"
+            )
+            raise AtError(msg)
+    detuning_elem = gen_detuning_elem(
+        split_ring,
+        qpx=qpx,
+        qpy=qpy,
+        detuning_coeff=detuning_coeff,
+        alphac=alphac,
+        orbit=o4[-1],
+        orbit6=o6[-1],
     )
-    return fastringnorad, fastringrad
+    if emity is None and emitx is None and espread is None:
+        qd_elem = gen_quantdiff_elem(split_ring.enable_6d(copy=True), orbit=o6[-1])
+    else:
+        params = split_ring.enable_6d(copy=True).envelope_parameters()
+        taux, tauy, tauz = params.Tau
+        if emitx is None:
+            emitx = params.emittances[0]
+        if emity is None:
+            emity = params.emittances[1]
+        if espread is None:
+            espread = params.sigma_e
+        qd_elem = SimpleQuantDiff(
+            "Diffusion",
+            betax=detuning_elem.BetaRad[0],
+            betay=detuning_elem.BetaRad[1],
+            emitx=emitx,
+            emity=emity,
+            espread=espread,
+            taux=taux * clight / ring.circumference,
+            tauy=tauy * clight / ring.circumference,
+            tauz=tauz * clight / ring.circumference,
+        )
+    fastring.append(detuning_elem)
+    fastring.append(qd_elem)
+    fastring = Lattice(fastring, **vars(ring))
+    if ring.radiation:
+        fastring.enable_6d()
+    else:
+        fastring.disable_6d()
+    return fastring
 
 
 def simple_ring(
